Update TdsipalyS3CalcCalendar.ino
ledcWrite moved under the WiFi delay on initial setup because it was too fast for the device and sometimes it wouldn't set the correct brightness when device booted.

There is now a complete working calculator where each operator calculates and displays the previous result and then waits for new operation, so we can do a 1+1+1 without the need of pressing in every operation. The operator is being displayed on screen so we know what has been pressed. Pressing a new operator after pressing another operator continues to display the correct result, along with the latest operator pressed.

The header of the calculator displays the WiFi SSID.

Code format fixes.